Reduce reliance on NMS modules (#1750)
* Move action bar messaging out of NMS modules using Adventure

* Move lightning strikes out of NMS modules using Bukkit API

* Use temporary build of adventure-platform-bukkit to restore actionbar support for 1.21.7+

Temporary build is hosted on GitHub Packages in this repository. This is only necessary because we are still supporting Bukkit/Spigot environments. If we targeted Paper and downstream forks then we could use Paper's native Adventure libraries.

* Move language check calls out of version modules where possible

* Refactor v1_19_R1 module

* Remove all NMS calls in IconMenuInventory

* Remove majority of EggIconService classes

Most icons are successfully grabbed with the API and don't need NMS classes

* Remove reliance on NMS for CustomInventory

May require override methods for 1.12. Tested 1.8.8, 1.16.5, 1.21.10