Add support for Copper Golems, oxidation, and waxing
Right click on a Copper Golem Pet with an Axe to strip a layer of oxidation.
Right click on a Copper Golem Pet with Honeycomb to prevent it from oxidizing.
Right click on a Copper Golem Pet with a Poppy to have it place the poppy on its head.
Right click a Copper Golem Pet with Shears to remove the Poppy from its head.

Adds two config options for Copper Golems in pet-config.yml:
CanOxidize: (default true), enables or disables oxidation of Copper Golem Pets
OxidationTime: (default 2400), time in ticks between each oxidation state

Copper Golem Pets do not turn into Statues, but could be possible in a future update